requirements.txt
Django==1.9.4
Markdown==2.6.6
Pillow==3.2.0
dj-database-url==0.4.0
dj-static==0.0.6
django-markdown-deux==1.0.5
django-markdownx==1.5
django-nested-inline==0.3.6
django-summernote==0.8.3
django-toolbelt==0.0.1
djangorestframework==3.3.3
gunicorn==19.4.5
markdown2==2.3.1
newrelic==2.60.0.46
psycopg2==2.6.1
requests==2.9.1
static3==0.6.1
whitenoise==3.0
